requirements.txt
dj_database_url==2.1.0
Django==5.0.4
djangorestframework==3.15.1
